third_party/blink/web_tests/external/wpt/html/canvas/offscreen/text/2d.text.direction.valid.worker-expected.txt

This is a testharness.js-based test.
[FAIL] 2d
  assert_equals: ctx.direction === 'inherit' (got ltr[string], expected inherit[string]) expected "inherit" but got "ltr"
Harness: the test ran to completion.
